Generate File & Beginnings Of Palabras
I have generated a JSON list of the 1000 most common words in Spanish,
along with their translations and ranks. I have done some hand massaging
of the data to normalize it, but I will need to do a lot more before
we are done, and I would also like to add some examples of using it in
a sentence, and also I'd like to add multiple translations for the words
which have multiple translations.